pyproject.toml
[tool.poetry]
name = "sylvain_eric_python"
version = "0.1.0"
description = ""
authors = ["sylvain <sylvain.ecuer@epita.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.1"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pylint = "^2.13.5"
SQLAlchemy = "^1.4.35"
pydantic = "^1.9.0"
pytest-mock-resources = "^2.2.7"
psycopg2-binary = "^2.9.5"
protobuf = "^4.21.12"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
grpclib = "^0.4.3"
jwt = "^1.3.1"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pre-commit = "^2.18.1"
black = "^22.3.0"
autoflake = "^1.4"
isort = "^5.10.1"
mypy = "^0.942"
rope = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 100
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY',
'LOCALFOLDER']
combine_as_imports = true

[tool.black]
line-length = 100

[tool.poetry.scripts]
start = "app:start"
lint = "app:lint"
